Error in vasp.method when multiple proc deleting same directory tree.

+++ vasp/methods.py
@@ -116,6 +116,9 @@ class RelaxCellShape(object):
                        .format(self.ediffg, self.vasp.ediff))
     ediffg *= 1.2 * float(len(structure.atoms))
 
+    is_mpi = False if comm == None else comm.size > 1
+    is_root = comm.rank == 0 if is_mpi else True
+
     # updates vasp as much as possible.
     if "set_relaxation" in kwargs: 
       warn("set_relaxation is deprecated. Please use relaxation.", DeprecationWarning)
@@ -211,7 +214,7 @@ class RelaxCellShape(object):
              )
     yield output
 
-    if output.success and (not keep_steps):
+    if output.success and (not keep_steps) and is_root:
       rmtree(join(outdir, "relax_cellshape"))
       rmtree(join(outdir, "relax_ions"))
 
@@ -246,7 +249,6 @@ class RelaxCellShape(object):
     from os.path import exists
     from ..opt import RelativeDirectory
 
-    is_root = True if comm == None else comm.rank == 0
     outdir = getcwd() if outdir == None else RelativeDirectory(outdir).path
     if not overwrite:
       extract = self.Extract(outdir, comm=None)
@@ -260,10 +262,8 @@ class RelaxCellShape(object):
     """ Just creates directory for debugging. """
     from ..opt.changedir import Changedir
 
-    is_root = True
-    if "comm" in kwargs:
-      comm = kwargs["comm"]
-      is_root = True if comm == None else comm.rank == 0
+    is_mpi = False if kwargs.get('comm', None) == None else kwargs['comm'].size > 1
+    is_root = kwargs['comm'].rank == 0 if is_mpi else True
     if is_root:
       # creates a file describing the relaxation parameters.
       with Changedir(kwargs["outdir"]) as pwd:
